# openapi.yaml
openapi: 3.0.0
info:
  title: API de Charadas
  description: Uma API para gerenciar e consultar charadas divertidas, com integração ao Firebase.
  version: "1.0"
  contact:
    name: Mayara

servers:
  - url: https://api-de-charadass.vercel.app/
    description: Ambiente em Produção

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Charada:
      type: object
      properties:
        id:
          type: integer
          example: 1
        pergunta:
          type: string
          example: "O que é, o que é?"
        resposta:
          type: string
          example: "É o que é."
    
    CharadaInput:
      type: object
      required:
        - pergunta
        - resposta
      properties:
        pergunta:
          type: string
        resposta:
          type: string

    Error:
      type: object
      properties:
        error:
          type: string

paths:
  /:
    get:
      summary: Informações da API
      responses:
        '200':
          description: Boas-vindas e versão
          content:
            application/json:
              example:
                api: "charadas"
                version: "1.0"
                author: "Mayara"

  /login:
    post:
      summary: Realiza login para obter o token de acesso
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                usuario:
                  type: string
                senha:
                  type: string
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              example:
                message: "Login realizado com sucesso!"
                token: "seu_jwt_aqui"
        '401':
          description: Usuário ou senha inválidos

  /charadas:
    get:
      summary: Retorna todas as charadas (Público)
      responses:
        '200':
          description: Lista de charadas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Charada'
    post:
      summary: Cadastra uma nova charada (Privado)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharadaInput'
      responses:
        '201':
          description: Criado com sucesso
        '400':
          description: Dados inválidos
        '401':
          description: Token ausente ou inválido

  /charadas/aleatoria:
    get:
      summary: Retorna uma charada aleatória (Público)
      responses:
        '200':
          description: Uma charada única
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Charada'

  /charadas/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Busca uma charada pelo ID (Público)
      responses:
        '200':
          description: Detalhes da charada
        '404':
          description: Charada não encontrada
    
    put:
      summary: Altera totalmente uma charada (Privado)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharadaInput'
      responses:
        '200':
          description: Alterado com sucesso
        '404':
          description: Não encontrado
    
    patch:
      summary: Altera parcialmente uma charada (Privado)
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                pergunta:
                  type: string
                resposta:
                  type: string
      responses:
        '200':
          description: Atualizado com sucesso
    
    delete:
      summary: Exclui uma charada (Privado)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Excluído com sucesso
        '404':
          description: Não encontrado